requirements.txt
airportsdata==20220107
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.3
et-xmlfile==1.1.0
furl==2.1.3
greenlet==1.1.2
idna==3.3
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jwcrypto==1.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.0
oauthlib==3.1.1
openpyxl==3.0.9
orderedmultidict==1.0.1
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
prettytable==0.7.2
prompt-toolkit==3.0.24
psycopg2==2.9.3
pyairports @ git+git://github.com/NICTA/pyairports@9699eae8c6b03c97239018dea92e01daa8594936
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-oauth2client==0.18.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.13.3
